Fixed some of the issues found with MSFT static code analysis (#185)
Most importantly, fixed issues with CCP log macro arguments not matching the format string as that now generates compile time errors on macOS

+++ trinity/Tr2BoneTransformBuffer.cpp
@@ -35,7 +35,7 @@ uint32_t Tr2BoneTransformBuffer::UploadTransforms( const Float4x3* data, uint32_
 	}
 	if( m_head < m_tail && m_head + matrixCount >= m_tail )
 	{
-		CCP_LOG( "Resizing bone transform buffer to %uKB", ( m_size * 2 * sizeof( Float4x3 ) ) / 1024 );
+		CCP_LOG( "Resizing bone transform buffer to %uKB", uint32_t( m_size * 2 * sizeof( Float4x3 ) ) / 1024 );
 		Resize( m_size * 2 );
 	}
